Try to make sure the translated code will not exceed 64KB
There was a hidden limit in the maximum size of the generated
code. Since the tb_jmp_offset fields are 16-bit wide, if the
offset of the jump instruction is larger than that, the value
will overflow and any tb_reset_jump() call would overwrite an
unrelated instruction. This took quite a while to track down.

So for now, just try to limit the maximum number of TCG uops
that we need to translate in one go. Notice that this partly
undoes the effects of bcc6ae14 (in the goldfish repo), which
changed the limit from 640 (in upstream qemu) to 2048 to work
around a buffer overflow when translating Neon instructions. The
value of 1024 tries to find a middle ground. The aforementioned
patch adds a check for the overflow and this patch adds a check
for the wraparound, so at least this cannot lead to silent
corruption, just an abort().